requirements.txt
Django==1.11.26
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.3
